@@ -23,7 +23,8 @@ class _CityScreenState extends State<CityScreen> {
             children: <Widget>[
               Align(
                 alignment: Alignment.topLeft,
-                child: FlatButton(
+                child: TextButton(
+                  style: kFlatButtonStyle,
                   onPressed: () {},
                   child: Icon(
                     Icons.arrow_back_ios,
@@ -35,7 +36,8 @@ class _CityScreenState extends State<CityScreen> {
                 padding: EdgeInsets.all(20.0),
                 child: null,
               ),
-              FlatButton(
+              TextButton(
+                style: kFlatButtonStyle,
                 onPressed: () {},
                 child: Text(
                   'Get Weather',
